lab9/shop-back/api/views.py

from rest_framework import status, viewsets
from rest_framework.decorators import action
from rest_framework.response import Response

from .models import Category, Product
from .serializers import CategorySerializer, ProductSerializer


class CategoryViewSet(viewsets.ModelViewSet):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer

    @action(detail=True, methods=["get"])
    def products(self, request, pk=None):
        category = self.get_object()
        serializer = ProductSerializer(category.products.all(), many=True)
        return Response(serializer.data)


class ProductViewSet(viewsets.ModelViewSet):
    queryset = Product.objects.all().order_by("-price")
    serializer_class = ProductSerializer

    def create(self, request,*args, **kwargs):
        if request.data.get("count", 0) < 0 or request.data.get("is_active") is False:
            return Response(
                {"detail": "error"},
                status=status.HTTP_400_BAD_REQUEST,
            )
        return super().create(request, *args, **kwargs)

