lab9/shop-back/requirements.txt
asgiref==3.11.1
djangorestframework==3.17.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3
